Refactor entities in sub modules
